Designed buttons Moved dashboardHeader into it's own component.
