+++ styles/Dashboard.module.css
@@ -1,0 +1,94 @@
+.main {
+  width: 100%;
+}
+
+.header {
+  width: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
+  padding: 60px;
+  padding-bottom: 0;
+}
+
+/* Buttons */
+.buttons {
+  display: flex;
+}
+
+.outline {
+  align-items: center;
+  background-image: linear-gradient(144deg, #6077ff, #cc75ff);
+  border: 0;
+  border-radius: 8px;
+  box-shadow: rgba(151, 65, 252, 0.2) 0 15px 30px -5px;
+  box-sizing: border-box;
+  color: #FFFFFF;
+  display: flex;
+  justify-content: center;
+  line-height: 1em;
+  max-width: 100%;
+  min-width: 140px;
+  padding: 3px;
+  text-decoration: none;
+  user-select: none;
+  -webkit-user-select: none;
+  touch-action: manipulation;
+  white-space: nowrap;
+  cursor: pointer;
+}
+
+.destroy {
+  background-image: linear-gradient(144deg, #cc75ff, #FFAABC);
+}
+
+.outline:active,
+.outline:hover {
+  outline: 0;
+}
+
+.outline span {
+  background-color: #3d1170;
+  padding: 16px 24px;
+  border-radius: 6px;
+  width: 100%;
+  height: 100%;
+  transition: 300ms;
+}
+
+.destroy span {
+  background-color: #4E050C;
+}
+
+.outline:hover span {
+  background: none;
+}
+
+@media (min-width: 768px) {
+  .outline {
+    font-size: 18px;
+    min-width: 150px;
+  }
+}
+
+/* Environment */
+.envDetails {
+  width: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  flex-direction: column;
+}
+
+.envContainer {
+  width: 75%;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  padding: 50px;
+}
+
+
+
+
