+++ styles/globals.css
@@ -1,4 +1,4 @@
-@import url('https://fonts.googleapis.com/css2?family=Poppins:wght@600&display=swap');
+@import url('https://fonts.googleapis.com/css2?family=Poppins:wght@400;500;600;800&display=swap');
 
 html,
 body {
@@ -6,6 +6,7 @@ body {
   margin: 0;
   font-family: 'Poppins', sans-serif;
   overflow-x: hidden;
+  color: #3d1170;
 }
 
 a {
@@ -31,6 +32,34 @@ footer {
   width: 100%;
 }
 
+h1 {
+  font-size: 3em;
+  text-shadow: #cc75ff -1px -1px;
+  font-weight: 800;
+}
+
+h2 {
+  font-weight: 600;
+  font-size: 2.5em;
+}
+
+h3 {
+  font-weight: 500;
+  font-size: 1.5em;
+}
+
+p, dd, li {
+  font-weight: 400;
+  font-size: 1.25em;
+}
+
+button {
+  padding: 15px;
+  margin: 10px;
+  border-radius: 10px;
+  font-family: 'Poppins', sans-serif;
+}
+
 @media (prefers-color-scheme: dark) {
   html {
     color-scheme: dark;
@@ -40,5 +69,39 @@ footer {
     background: black;
   }
 }
+
+/* env Stack */
+.stacks {
+  display: grid;
+  grid-template-areas:
+    "head head"
+    "env  service";
+  width: 75%;
+  column-gap: 4rem;
+}
+
+.stacks h2 {
+  grid-area: head;
+  text-align: center;
+}
+
+.stacks .env {
+  grid-area: env;
+}
+
+.stacks .service {
+  grid-area: service;
+}
+
+dl div {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+dd {
+  margin-inline-start: 0;
+}
+
